requirements.txt
faiss-cpu==1.7.3
langchain==0.0.146
openai==0.27.4
pandas==1.5.3
python-dotenv==1.0.0
requests==2.28.2
streamlit==1.21.0
streamlit-chat==0.0.2.2
pypdf==3.8.1
pdfminer.six==20221105
tiktoken==0.3.3

# supabase==1.0.3
qdrant_client==1.1.7
protobuf==4.23.1
